// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/radiology/shared/DTOs/file-upload.dto.ts
export class FileUpload_DTO {
    // PatientDetails: Patient;
    // FileDetails: PatientFilesModel
    PatientId: number = 0;
    ImagingReportId: number = 0;
    ImagingItemName: string;
    ImagingRequisitionId: string;
    PatientFileId: number = 0;
    PatientCode: string;
    OrderStatus: string;
    PerformerName: string;
    PerformerId: number;

}